requirements.txt
annotated-types==0.7.0
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
pydantic==2.11.9
pydantic_core==2.33.2
PyYAML==6.0.2
requests==2.32.5
-e git+https://github.com/AGASTRONICS/secploy-python-sdk.git@448692f72d13519e2e8548e74e2904b50d9a0703#egg=secploy
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
